chore: refactor `utils`
