+++ tsconfig.json
@@ -6,7 +6,7 @@
     "target": "es2019",
     "moduleResolution": "node",
     "strictBindCallApply": true,
-    "allowSyntheticDefaultImports": true,
+    "allowSyntheticDefaultImports": false,
     "forceConsistentCasingInFileNames": true,
     "strictFunctionTypes": true,
     "resolveJsonModule": true,
